Restructured "image capture"
Beginning the process of restructuring functions to make them more
modular. Altered the name of image capture function to drawImg,
reflecting what it really does. Now accepts a video and canvas element,
drawing the current image in the video on the canvas element. Passed
the output canvas and the video element to this function to test.
Output canvas shows image from webcam.