Added start camera stream function
Added start camera stream function. Starts the webcam video stream.
Works with the getUserMedia API. Other APIs might need different code
to perform this function. Gum has been put in filename to reflect that
it works with gum. If different code is needed for a different API, the
file need only be substituted for another file.