Calc bw threshold changed to most colour val
Changed black/white threshold function name to most colour value.
Reflects function’s action rather than its purpose. It is used to
calculate the black and white threshold. Its action is to determine the
most occurring colour value in a picture. The action is an image
processing function; the purpose is a MagText operation.