threshOffset/bwThreshold names changed
Variable threshOffset has been renamed to txtColValDist. The offset is
actually the distance from the most common background colour to the
start of the text. The variable has been renamed to this effect. A call
to binary picture shows that it is being correctly received as an
argument. Similarly, bwThreshold has been changed to bgMostColVal,
since it represents the most common colour value in the background.
Call to binary picture and mostColVal() shows that the change has been
successfully implemented.