Text distance variable name made more precise
Updated text distance variable name to reflect that it is a minimum
rather than a maximum or value in a range. Now need to form the adjust
content function. Will include the commented-out sharpening function
just before the quantisation function to sharpen out of focus pictures.
This will be used at a later stage during testing