Button begins web app
Web app is now started by clicking the Capture button. Later, this
button will be modified to perform a reload of the page. For now, its
use is to
suspend execution until input values are specified. The whole web app
can be used for further development later down the track.